+++ Resources/Model/NotificationBody.php
@@ -6,7 +6,7 @@ namespace Reliefapps\NotificationBundle\Resources\Model;
 class NotificationBody
 {
     const PAYLOAD_ARRAY_ANDROID  = 0;
-    const PAYLOAD_JSON_IOS      = 1;
+    const PAYLOAD_JSON_IOS       = 1;
 
     /**
      *  @var String
@@ -27,10 +27,22 @@ class NotificationBody
     private $badge;
 
     /**
+     *  @var Integer
+     *  Id of the notification (to distinguish them)
+     */
+    private $notId;
+
+    /**
      *  @var Array
-     *  Color or the led for android
+     *  Color of the led for android
      */
     private $ledColor;
+
+    /**
+     *  @var String
+     *  Path to the large icon
+     */
+    private $image;
 
     /**
      *  @var String
@@ -44,32 +56,41 @@ class NotificationBody
      */
     private $actions;
 
+    /**
+     *  @var Array
+     *  List of fields to add to the notification
+     */
+    private $additionalFields;
+
     public function __construct()
     {
-        $this->title    = null;
-        $this->body     = null;
-        $this->badge    = null;
-        $this->ledColor = null;
-        $this->category = null;
-        $this->actions  = null;
+        $this->title              = null;
+        $this->body               = null;
+        $this->badge              = null;
+        $this->notId              = null;
+        $this->ledColor           = null;
+        $this->image              = null;
+        $this->imageType          = null;
+        $this->category           = null;
+        $this->actions            = null;
+        $this->additionalFields   = null;
     }
 
     /**
      *  Get Payload for a given device
      *
      *  @param Integer Payload Type (see constants)
-     *  @param Array(Associative arrays) Additional fields
      *
      *  @return String Json payload
      */
-    public function getPayload($payload_type, $additionalFields)
+    public function getPayload($payload_type)
     {
         switch ($payload_type) {
             case self::PAYLOAD_JSON_IOS:
-                return $this->getiOSPayload($additionalFields);
+                return $this->getiOSPayload();
                 break;
             case self::PAYLOAD_ARRAY_ANDROID:
-                return $this->getAndroidPayload($additionalFields);
+                return $this->getAndroidPayload();
                 break;
 
             default:
@@ -78,7 +99,7 @@ class NotificationBody
         }
     }
 
-    private function getiOSPayload($additionalFields)
+    private function getiOSPayload()
     {
         $payload = array(
                 "aps" => array(
@@ -94,17 +115,19 @@ class NotificationBody
         if($this->getCategory()){
             $payload["aps"]["category"] = $this->getCategory();
         }
-
-        foreach($additionalFields as $additionalField){
-            if(array_key_exists("key", $additionalField) && array_key_exists("value", $additionalField)){
-                $payload["aps"][$additionalField["key"]] = $additionalField["value"];
+        if($this->getAdditionalFields()){
+            $additionalFields = $this->getAdditionalFields();
+            foreach($additionalFields as $additionalField){
+                if(array_key_exists("key", $additionalField) && array_key_exists("value", $additionalField)){
+                    $payload["aps"][$additionalField["key"]] = $additionalField["value"];
+                }
             }
         }
 
         return json_encode($payload);
     }
 
-    private function getAndroidPayload($additionalFields)
+    private function getAndroidPayload()
     {
         $payload = array();
         if($this->getTitle()){
@@ -113,16 +136,27 @@ class NotificationBody
         if($this->getBody()){
             $payload["message"] = $this->getBody();
         }
-        if($this->getActions()){
-            $payload["actions"] = $this->getActions();
+        if($this->getNotId()){
+            $payload["notId"] = $this->getNotId();
         }
         if($this->getLedColor()){
             $payload["ledColor"] = $this->getLedColor();
         }
-
-        foreach($additionalFields as $additionalField){
-            if(array_key_exists("key", $additionalField) && array_key_exists("value", $additionalField)){
-                $payload[$additionalField["key"]] = $additionalField["value"];
+        if($this->getImage()){
+            $payload["image"] = $this->getImage();
+        }
+        if($this->getImageType()){
+            $payload["image-type"] = $this->getImageType();
+        }
+        if($this->getActions()){
+            $payload["actions"] = $this->getActions();
+        }
+        if($this->getAdditionalFields()){
+            $additionalFields = $this->getAdditionalFields();
+            foreach($additionalFields as $additionalField){
+                if(array_key_exists("key", $additionalField) && array_key_exists("value", $additionalField)){
+                    $payload[$additionalField["key"]] = $additionalField["value"];
+                }
             }
         }
 
@@ -202,6 +236,30 @@ class NotificationBody
     }
 
     /**
+     * Get the value of notId
+     *
+     * @return Integer
+     */
+    public function getNotId()
+    {
+        return $this->notId;
+    }
+
+    /**
+     * Set the value of notId
+     *
+     * @param Integer notId
+     *
+     * @return self
+     */
+    public function setNotId($notId)
+    {
+        $this->notId = $notId;
+
+        return $this;
+    }
+
+    /**
      * Get the value of Led Color
      *
      * @return Array
@@ -221,6 +279,54 @@ class NotificationBody
     public function setLedColor($ledColor)
     {
         $this->ledColor = $ledColor;
+
+        return $this;
+    }
+
+    /**
+     * Get the value of Image
+     *
+     * @return string
+     */
+    public function getImage()
+    {
+        return $this->image;
+    }
+
+    /**
+     * Set the value of Image
+     *
+     * @param string image
+     *
+     * @return self
+     */
+    public function setImage($image)
+    {
+        $this->image = $image;
+
+        return $this;
+    }
+
+    /**
+     * Get the value of imageType
+     *
+     * @return string
+     */
+    public function getImageType()
+    {
+        return $this->imageType;
+    }
+
+    /**
+     * Set the value of imageType
+     *
+     * @param string imageType
+     *
+     * @return self
+     */
+    public function setImageType($imageType)
+    {
+        $this->imageType = $imageType;
 
         return $this;
     }
@@ -262,7 +368,7 @@ class NotificationBody
     /**
      * Add an action
      *
-     * @return Array
+     * @return self
      */
     public function addAction($action)
     {
@@ -281,6 +387,42 @@ class NotificationBody
     public function setActions($actions)
     {
         $this->actions = $actions;
+
+        return $this;
+    }
+
+    /**
+     * Get the value of additionalFields
+     *
+     * @return Array
+     */
+    public function getAdditionalFields()
+    {
+        return $this->additionalFields;
+    }
+
+    /**
+     * Add a field
+     *
+     * @return self
+     */
+    public function addAdditionalField($additionalField)
+    {
+        $this->additionalFields[] = $additionalField;
+
+        return $this;
+    }
+
+    /**
+     * Set the value of additionalFields
+     *
+     * @param Array additionalFields
+     *
+     * @return self
+     */
+    public function setAdditionalFields($additionalFields)
+    {
+        $this->additionalFields = $additionalFields;
 
         return $this;
     }